Don't wrap pci_cleanup in exception handler
There's no point since it will either work if we're in the correct state (i.e.
we have pci_alloc'd an access object) or it will fail (i.e. we've called it
twice). However, this failure mode is not handle-able since it will cause the
process to die with a double-free.

Signed-off-by: Si Beaumont <simon.beaumont@citrix.com>